backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
chromadb
python-dotenv
numpy
onnxruntime
